+++ feelpp/tools/mesh/mesh_partitioner_hypercube.cpp
@@ -21,7 +21,8 @@
 //! License along with this library; if not, write to the Free Software
 //! Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301  USA
 //!
-#include <mesh_partitioner.hpp>
+
+#include <mesh_partitioner_impl.hpp>
 
 namespace Feel {
