+++ feelpp/tools/mesh/partitioner.adoc
@@ -57,18 +57,20 @@ We generate a mesh partitioned on 4 cores with the following command
 
 [source,shell]
 ----
-feelpp_mesh_partitioner --part 4  --ifile src/feelpp/data/gmsh/primitives/torus.geo --ofile torus
+feelpp_mesh_partitioner --part 4  --ifile /usr/share/feelpp/data/geo/torus.geo --odir torus_partitioning   
 ----
 
 You should have in the current directory 4 files
 
 [source,shell]
 ----
-ls torus*
-torus.geo  torus.h5  torus.json  torus.msh
+ls torus_partitioning/
+torus_p4.h5  torus_p4.json
 ----
 
-Now the file `torus.json` can be loaded in a Feel++ application distributed on 4 cores.
+Now the file `torus_p4.json` can be loaded in a Feel++ application distributed on 4 cores.
+
+NOTE: Others file generated by the application (msh file from geo, log, ...) are located into the app directory (which is controled by `directory` option)
 
 === Generating multiple mesh partitioning
 
@@ -77,21 +79,20 @@ Often we are interested in a set of partitioned meshes in order to do a speed-up
 
 [source,shell]
 ----
-feelpp_mesh_partitioner --part 2 4 8 16 32  --ifile src/feelpp/data/gmsh/primitives/torus.geo --odir torus-mesh
+feelpp_mesh_partitioner --part 2 4 8 16 32  --ifile  /usr/share/feelpp/data/geo/torus.geo --odir torus_partitioning   
 ----
 
 You should have in the  directory `torus-mesh` (thanks to the `odir` option) 5 partitioned meshes
 
 [source,shell]
 ----
-ls torus*
-torus_p16.h5    torus_p2.h5    torus_p32.h5    torus_p4.h5    torus_p8.h5
-torus_p16.json  torus_p2.json  torus_p32.json  torus_p4.json  torus_p8.json
+ls torus_partitioning
+torus_p16.h5  torus_p16.json  torus_p2.h5  torus_p2.json  torus_p32.h5  torus_p32.json  torus_p4.h5  torus_p4.json  torus_p8.h5  torus_p8.json
 ----
 
 NOTE: The mesh filenames contain the partition information.
 
-NOTE: If you want a different prefix than `torus`, like `torus-coarse`, you can use the `--ifile torus-coarse` option and you will have a set of files named `torus-coarse_p*.*` in the `torus-mesh` directory.
+NOTE: If you want a different prefix than `torus`, like `torus-coarse`, you can use the `--ofile torus-coarse` option and you will have a set of files named `torus-coarse_p*.*` in the `torus_partitioning` directory.
 
 === Generating a mesh partitioning by markers
 1) partitioning of each marker present in the mesh :
